feat!: use keccak256 for codehash
BREAKING CHANGE: Previously, codehash was calculated using poseidon.
So the state root since this commit will be produced differently.
This is because we are implementing Type2 zkEVM. But at this moment,
I am not sure why scroll team store poseidon codehash in account.
Maybe we can revert this commit after we know the reason and it makes
sense.

See belows for details.

- https://eips.ethereum.org/EIPS/eip-1052
- https://github.com/scroll-tech/go-ethereum/pull/188