feat(core): enable marshaling depositTxWithNonce
While implementing https://github.com/kroma-network/kroma/issues/17,
I faced the issue that a deposit transaction is transformed to
a deposit transaction with nonce during unmarshaling from rpc transaction.
The problem is marshaling of a transaction with nonce omits some of
properties this leads to several unexpected results.

This isn't meant to be a good solution for the regarding issue above.
For example, `SourceHash()` of a deposit transaction with nonce doesn't
return what we expect as a deposit transaction does.

Still, this is a enough change for me to get public input from output
root for ZK fault proof.